Conflicts:
	src/org/enderstone/server/packet/Packet.java
	src/org/enderstone/server/regions/EnderChunk.java

+++ src/org/enderstone/server/packet/Packet.java
@@ -325,14 +325,14 @@ public abstract class Packet {
 
 	public static Location readLocation(ByteBuf buf) {
 		long value = buf.readLong();
-		double x = value >> 38;
-		double y = value << 26 >> 52;
-		double z = value << 38 >> 38;
-		return new Location("", x, y, z, 0F, 0F);
+		long x = value >> 38;
+		long y = value << 26 >> 52;
+		long z = value << 38 >> 38;
+		return new Location("", x,y,z,0F,0F);
 	}
-
-	public static void writeLocation(Location loc, ByteBuf buf) {
-		buf.writeLong((loc.getBlockX() & 0x3FFFFFF) << 38 | (loc.getBlockY() & 0xFFF) << 26 | (loc.getBlockZ() & 0x3FFFFFF));
+	
+	public static void writeLocation(Location loc , ByteBuf buf){
+		buf.writeLong(((long)loc.getBlockX() & 0x3FFFFFF) << 38 | ((long)loc.getBlockY() & 0xFFF) << 26 | (loc.getBlockZ() & 0x3FFFFFF));
 	}
 
 	public static int getLocationSize() {
